Add "exempt" rating and wallet profiles.
Wallet profiles represent the type of use-case and audience that a wallet
is aiming for. For example, a wallet may be focusing on peer-to-peer payments.
Such wallets should not be rated about things like browser integration
standards adherence, because they do not aim to integrate with browsers in
this manner. They target a specific use-case and not all of them.

Wallets can now be classified by their profile, which can then be used to
exempt them from certain attributes, hence the need to introduce the "exempt"
rating.

This commit also refactors YES/NO ratings into PASS/FAIL which are more common
terms for them.